adds the new PROGMEMSTRING feature
A PROGMEMSTRING enables the placement of string constants in the text segment,
allowing the reduction of needed RAM on some processor architectures like the
AVRs.

Especially in the embedded field the RAM is often a constrained ressource.
Enabling the placement of constant strings in e.g. the FLASH memory allows
reducing the needed RAM. For the AVR platform, the logging framework supports
such a feature by using the PROGMEMSTRING macro. This macro places the string
constant in the none volatile memory, creates a temporarly variable that
abstracts the access to the given string, and is used by the logging framework.
On architecture without the need for such a feature, or on those not yet
supported, the macro substituation results in the string constant. Thus, the
PROGMEMSTRING macro can be used always.