fixes problem with program memory string
The program memory string feature works not correctly on AVRs if a
compiler newer than 4.2 is used. Those compilers have a bug within the
attribute progmem interpretation as well as the prog_char typedef. With
the new implemented scheme it works with the newer compiler versions,
too. Furthermore, the program memory string is usable on all systems as
type, and its behavior is that of a pointer to character array, allowing
dereferencing, pointer arithmetic, and indexing.