rework on PROGMEMSTRING feature
On those systems not needing such a feature, the macro substitutes to
the string constant, only. Nevertheless, the ProgmemString is usable  as
well on such systems. Furthermore, some glitches with not well behaving
compilers are workaround. One thing is the placement of the logging
strings within special section beginning with ".progmem.logging-cpp.".
This allows selective removing of messages through the linker by garbage
collecting unused sections. Doing not so, leading to have all log
messages in one section. Disabling some levels would work, but despite
this the messages of disabled level would be part of the executable, but
this needs to be avoided.